Fixed a bug with poll & pollend and changed examples.
Fixed a bug where if 11+ arguments were provided to $poll, the bot would break.
The bot also would break on $pollend if 11+ polls were created.

Now if there are 11+ args provided to $poll, anything higher than 10 would be ignored.
Same with the number of polls for $pollend.

Changed the examples to be in a table.